requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
jwt==1.3.1
mysqlclient==2.2.0
Pillow==10.0.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
